Separate components from reducres
