+++ src/components/categories.jsx
@@ -1,0 +1,7 @@
+const Categories = () => (
+  <div className="categories">
+    <h2>Under construction</h2>
+  </div>
+);
+
+export default Categories;
