--- src/redux/categories/categories.js
@@ -1,7 +1,0 @@
-const Categories = () => (
-  <div className="categories">
-    <h2>Under construction</h2>
-  </div>
-);
-
-export default Categories;
