+++ src/redux/configureStore.js
@@ -9,4 +9,6 @@ const reducer = combineReducers({
 const store = createStore(
   reducer,
   applyMiddleware(logger)
-)
+);
+
+export default store;
